# improve clickhouse_query_perf.py script: