--- a/tools/torchci/requirements.txt
+++ b/tools/torchci/requirements.txt
@@ -3,3 +3,6 @@ requests
 boto3==1.19.12
 pytest==7.2.0
 clickhouse-connect==0.8.14
+prettytable
+tqdm
+python-dotenv