Changes:
* add correct dependencies
* by default `--head` now uses unstaged query; now:
   *  when running `--perf` without specifying `--base` it will just run tests for unstaged changes
   * when running `--results` without specifying `--base` and `--head`, it will compare committed version vs local changes
* add `--verbose` flag to print results

### Testing

```
./clickhouse_query_perf.py --query weekly_force_merge_stats  --perf --times 3 --base HEAD
Using unstaged changes for --head
Gathering perf stats for: weekly_force_merge_stats
Num tests: 2
Num times: 3
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 3/3 [00:08<00:00,  2.85s/it]
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 3/3 [00:02<00:00,  1.25it/s]
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 3/3 [00:06<00:00,  2.31s/it]
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 3/3 [00:02<00:00,  1.41it/s]
+------+----------+-----------+-------------+---------------+------------+-----------+------------+--------------+
| Test | Avg Time | Base Time | Time Change | % Time Change |  Avg Mem   |  Base Mem | Mem Change | % Mem Change |
+------+----------+-----------+-------------+---------------+------------+-----------+------------+--------------+
|  0   |   2395   |    632    |     1763    |      279      | 4096500724 |  87111215 | 4009389509 |     4603     |
|  1   |   2148   |    595    |     1553    |      261      | 4244083446 | 114181482 | 4129901964 |     3617     |
+------+----------+-----------+-------------+---------------+------------+-----------+------------+--------------+

```



```
 ./clickhouse_query_perf.py --query weekly_force_merge_stats  --results --base HEAD
Using unstaged changes for --head
Comparing results for query: weekly_force_merge_stats
Num tests: 2
Head: unstaged
 Base: HEAD
Results for test 0 match
Results for test 1 match
```


```
./clickhouse_query_perf.py --query weekly_force_merge_stats  --results --base HEAD --verbose
Using unstaged changes for --head
Comparing results for query: weekly_force_merge_stats
Num tests: 2
Head: unstaged
 Base: HEAD

=== Test 0 Results (first 5 rows) ===
Row 0: {
  "granularity_bucket": "2025-03-03",
  "metric": 24.63768115942029,
  "name": "All Force Merges"
}
Row 1: {
  "granularity_bucket": "2025-03-03",
  "metric": 14.009661835748792,
  "name": "From Failures"
}
Row 2: {
  "granularity_bucket": "2025-03-03",
  "metric": 10.628019323671497,
  "name": "From Impatience"
}
Row 3: {
  "granularity_bucket": "2025-02-24",
  "metric": 31.25,
  "name": "All Force Merges"
}
Row 4: {
  "granularity_bucket": "2025-02-24",
  "metric": 9.375,
  "name": "From Failures"
}
Total rows: 6
==================================================
Results for test 0 match

=== Test 1 Results (first 5 rows) ===
Row 0: {
  "granularity_bucket": "2025-02-24",
  "metric": 27.134146341463413,
  "name": "All Force Merges"
}
Row 1: {
  "granularity_bucket": "2025-02-24",
  "metric": 14.634146341463415,
  "name": "From Failures"
}
Row 2: {
  "granularity_bucket": "2025-02-24",
  "metric": 12.5,
  "name": "From Impatience"
}
Row 3: {
  "granularity_bucket": "2025-02-17",
  "metric": 34.23913043478261,
  "name": "All Force Merges"
}
Row 4: {
  "granularity_bucket": "2025-02-17",
  "metric": 20.108695652173914,
  "name": "From Failures"
}
Total rows: 15
==================================================
Results for test 1 match
```